Refactored into a base class and two subclasses
