+++ apps/dav/lib/Connector/Sabre/ServerFactory.php
@@ -28,8 +28,9 @@
 
 namespace OCA\DAV\Connector\Sabre;
 
+use OCA\DAV\DAV\FileCustomPropertiesBackend;
+use OCA\DAV\DAV\MiscCustomPropertiesBackend;
 use OCA\DAV\Files\BrowserErrorPagePlugin;
-use OCA\DAV\DAV\CustomPropertiesBackend;
 use OCP\Files\Mount\IMountManager;
 use OCP\IConfig;
 use OCP\IDBConnection;
@@ -178,16 +179,34 @@ class ServerFactory {
 					\OC::$server->getGroupManager(),
 					$userFolder
 				));
+
 				// custom properties plugin must be the last one
-				$server->addPlugin(
-					new \Sabre\DAV\PropertyStorage\Plugin(
-						new CustomPropertiesBackend(
-							$objectTree,
-							$this->databaseConnection,
-							$this->userSession->getUser()
-						)
+				$filePropertiesPlugin = new \Sabre\DAV\PropertyStorage\Plugin(
+					new FileCustomPropertiesBackend(
+						$objectTree,
+						$this->databaseConnection,
+						$this->userSession->getUser()
 					)
 				);
+				$filePropertiesPlugin->pathFilter = function($path) {
+					$pathInfo = $this->request->getPathInfo();
+					// matches '/webdav', '/webdav/', '/webdav/adcd' but not '/webdavabcd'
+					return preg_match('!^/webdav(/|\z)!', $pathInfo);
+				};
+				$server->addPlugin($filePropertiesPlugin);
+
+				$miscPropertiesPlugin = new \Sabre\DAV\PropertyStorage\Plugin(
+					new MiscCustomPropertiesBackend(
+						$objectTree,
+						$this->databaseConnection,
+						$this->userSession->getUser()
+					)
+				);
+				$miscPropertiesPlugin->pathFilter = function($path) {
+					$pathInfo = $this->request->getPathInfo();
+					return preg_match('!^/webdav(/|\z)!', $pathInfo);
+				};
+				$server->addPlugin($miscPropertiesPlugin);
 			}
 			$server->addPlugin(new \OCA\DAV\Connector\Sabre\CopyEtagHeaderPlugin());
 		}, 30); // priority 30: after auth (10) and acl(20), before lock(50) and handling the request
