+++ apps/dav/lib/DAV/FileCustomPropertiesBackend.php
@@ -1,0 +1,196 @@
+<?php
+/**
+ * @author Thomas Müller <thomas.mueller@tmit.eu>
+ * @author Vincent Petry <pvince81@owncloud.com>
+ * @author Viktar Dubiniuk <dubiniuk@owncloud.com>
+ *
+ * @copyright Copyright (c) 2017, ownCloud GmbH
+ * @license AGPL-3.0
+ *
+ * This code is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License, version 3,
+ * as published by the Free Software Foundation.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License, version 3,
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>
+ *
+ */
+
+namespace OCA\DAV\DAV;
+
+use Doctrine\DBAL\Connection;
+use OCA\DAV\Connector\Sabre\Directory;
+use Sabre\DAV\INode;
+
+class FileCustomPropertiesBackend extends AbstractCustomPropertiesBackend {
+
+	const SELECT_BY_ID_STMT = 'SELECT * FROM `*PREFIX*properties` WHERE `fileid` = ?';
+	const INSERT_BY_ID_STMT = 'INSERT INTO `*PREFIX*properties`'
+	. ' (`fileid`,`propertyname`,`propertyvalue`) VALUES(?,?,?)';
+	const UPDATE_BY_ID_AND_NAME_STMT = 'UPDATE `*PREFIX*properties`'
+	. ' SET `propertyvalue` = ? WHERE `fileid` = ? AND `propertyname` = ?';
+	const DELETE_BY_ID_STMT = 'DELETE FROM `*PREFIX*properties` WHERE `fileid` = ?';
+	const DELETE_BY_ID_AND_NAME_STMT = 'DELETE FROM `*PREFIX*properties`'
+	. ' WHERE `fileid` = ? AND `propertyname` = ?';
+
+	/**
+	 * This method is called after a node is deleted.
+	 *
+	 * @param string $path path of node for which to delete properties
+	 */
+	public function delete($path) {
+		$node = $this->getNodeForPath($path);
+		if (is_null($node)) {
+			return;
+		}
+
+		$fileId = $node->getId();
+		$statement = $this->connection->prepare(self::DELETE_BY_ID_STMT);
+		$statement->execute([$fileId]);
+		$this->offsetUnset($fileId);
+		$statement->closeCursor();
+	}
+
+	/**
+	 * This method is called after a successful MOVE
+	 *
+	 * @param string $source
+	 * @param string $destination
+	 *
+	 * @return void
+	 */
+	public function move($source, $destination) {
+		// Part of interface. We don't care about move because it doesn't affect fileId
+	}
+
+	/**
+	 * @inheritdoc
+	 */
+	protected function getProperties($path, INode $node, array $requestedProperties) {
+		$fileId = $node->getId();
+		if (is_null($this->offsetGet($fileId))) {
+			// TODO: chunking if more than 1000 properties
+			$sql = self::SELECT_BY_ID_STMT;
+			$whereValues = [$fileId];
+			$whereTypes = [null];
+
+			if (!empty($requestedProperties)) {
+				// request only a subset
+				$sql .= ' AND `propertyname` in (?)';
+				$whereValues[] = $requestedProperties;
+				$whereTypes[] = Connection::PARAM_STR_ARRAY;
+			}
+
+			$props = $this->fetchProperties($sql, $whereValues, $whereTypes);
+			$this->offsetSet($fileId, $props);
+		}
+		return $this->offsetGet($fileId);
+	}
+
+	/**
+	 * @inheritdoc
+	 */
+	protected function updateProperties($path, INode $node, $changedProperties) {
+		$existingProperties = $this->getProperties($path, $node, []);
+		$fileId = $node->getId();
+		$deleteStatement = self::DELETE_BY_ID_AND_NAME_STMT;
+		$insertStatement = self::INSERT_BY_ID_STMT;
+		$updateStatement = self::UPDATE_BY_ID_AND_NAME_STMT;
+
+		// TODO: use "insert or update" strategy ?
+		$this->connection->beginTransaction();
+		foreach ($changedProperties as $propertyName => $propertyValue) {
+			$propertyExists = array_key_exists($propertyName, $existingProperties);
+			// If it was null, we need to delete the property
+			if (is_null($propertyValue)) {
+				if ($propertyExists) {
+					$this->connection->executeUpdate($deleteStatement,
+						[
+							$fileId,
+							$propertyName
+						]
+					);
+				}
+			} else {
+				if (!$propertyExists) {
+					$this->connection->executeUpdate($insertStatement,
+						[
+							$fileId,
+							$propertyName,
+							$propertyValue
+						]
+					);
+				} else {
+					$this->connection->executeUpdate($updateStatement,
+						[
+							$propertyValue,
+							$fileId,
+							$propertyName
+						]
+					);
+				}
+			}
+		}
+
+		$this->connection->commit();
+		$this->offsetUnset($fileId);
+
+		return true;
+	}
+
+	/**
+	 * Bulk load properties for directory children
+	 *
+	 * @param INode $node
+	 * @param array $requestedProperties requested properties
+	 *
+	 * @return void
+	 */
+	protected function loadChildrenProperties(INode $node, $requestedProperties) {
+		// note: pre-fetching only supported for depth <= 1
+		if (!($node instanceof Directory)){
+			return;
+		}
+
+		$fileId = $node->getId();
+		if (!is_null($this->offsetGet($fileId))) {
+			// we already loaded them at some point
+			return;
+		}
+
+		$childNodes = $node->getChildren();
+		$childrenIds = [];
+		// pre-fill cache
+		foreach ($childNodes as $childNode) {
+			$childId = $childNode->getId();
+			if ($childId) {
+				$childrenIds[] = $childId;
+				$this->offsetSet($childId, []);
+			}
+		}
+
+		// TODO: use query builder
+		$sql = 'SELECT * FROM `*PREFIX*properties` WHERE `fileid` IN (?)';
+		$sql .= ' AND `propertyname` in (?) ORDER BY `propertyname`';
+
+		$result = $this->connection->executeQuery(
+			$sql,
+			[$childrenIds, $requestedProperties],
+			[Connection::PARAM_STR_ARRAY, Connection::PARAM_STR_ARRAY]
+		);
+
+		$props = [];
+		while ($row = $result->fetch()) {
+			$props[$row['propertyname']] = $row['propertyvalue'];
+			$this->offsetSet($row['fileid'], $props);
+		}
+
+		$result->closeCursor();
+	}
+
+}
